Split Chattia UI from worker mesh
