+++ worker/edge-mesh/src/index.ts
@@ -2,6 +2,6 @@ import { jsonResponse } from "./../workers/shared/protocol";
 
 export default {
   async fetch(): Promise<Response> {
-    return jsonResponse({ message: "Ultimate chatbot root worker placeholder" });
+    return jsonResponse({ message: "Edge worker mesh root placeholder" });
   },
 };
