+++ worker/edge-mesh/workers/layer3-thorium/src/index.ts
