+++ worker/edge-mesh/workers/layer4-tiny-ml/src/index.ts
