+++ worker/edge-mesh/workers/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ultimate-chatbot-workers",
+  "name": "edge-worker-mesh",
   "version": "0.1.0",
   "private": true,
   "scripts": {
