+++ worker/edge-mesh/workers/shared/protocol.ts
