+++ worker/edge-mesh/wrangler.toml
@@ -1,4 +1,4 @@
-name = "ultimate-chatbot"
+name = "edge-worker-mesh"
 main = "src/index.ts"
 compatibility_date = "2024-05-01"
 
@@ -17,37 +17,37 @@ preview_id = ""
 logs = { enabled = true }
 
 [env.layer1]
-name = "ultimate-layer1-ui"
+name = "edge-layer1-ui"
 main = "workers/layer1-ui/src/index.ts"
 route = "https://layer1.example.com/*"
 
 [env.layer2]
-name = "ultimate-layer2-firewall"
+name = "edge-layer2-firewall"
 main = "workers/layer2-firewall/src/index.ts"
 route = "https://layer2.example.com/*"
 
 [env.layer3]
-name = "ultimate-layer3-thorium"
+name = "edge-layer3-thorium"
 main = "workers/layer3-thorium/src/index.ts"
 route = "https://layer3.example.com/*"
 
 [env.layer4]
-name = "ultimate-layer4-tiny-ml"
+name = "edge-layer4-tiny-ml"
 main = "workers/layer4-tiny-ml/src/index.ts"
 route = "https://layer4.example.com/*"
 
 [env.layer5]
-name = "ultimate-layer5-tiny-llm"
+name = "edge-layer5-tiny-llm"
 main = "workers/layer5-tiny-llm/src/index.ts"
 route = "https://layer5.example.com/*"
 
 [env.layer6]
-name = "ultimate-layer6-orchestrator"
+name = "edge-layer6-orchestrator"
 main = "workers/layer6-orchestrator/src/index.ts"
 route = "https://layer6.example.com/*"
 
 [env.layer7]
-name = "ultimate-layer7-backbone"
+name = "edge-layer7-backbone"
 main = "workers/layer7-backbone/src/index.ts"
 route = "https://layer7.example.com/*"
 
@@ -59,4 +59,4 @@ local_protocol = "http"
 # Service bindings can be declared per environment, e.g.
 # [[env.layer2.services]]
 # binding = "FIREWALL"
-# service = "ultimate-layer2-firewall"
+# service = "edge-layer2-firewall"
