+++ worker/layers/layer1-ui/README.md
@@ -12,7 +12,7 @@ interact with the orchestrator through the Layer 1 EdgeMesh Worker.
   provenance when calling the firewall and orchestrator layers.
 
 ## Setup Checklist
-1. Deploy the Worker at `infrastructure/edge-mesh/workers/layer1-ui` via
+1. Deploy the Worker at `edge-mesh/workers/layer1-ui` via
    Wrangler.
 2. Configure Pages or a static hosting bucket to serve your UI assets.
 3. Bind `SESSION_KV` for session storage, `FIREWALL_SERVICE` for Zero Trust
@@ -21,14 +21,14 @@ interact with the orchestrator through the Layer 1 EdgeMesh Worker.
 4. Document UI routes in `openapi.yaml` (create in this directory).
 
 ## Embed Widget & Multi-Tenant Controls
-- The Worker serves `/embed.js`, a drop-in script that renders a floating chat
-  UI inside any host page.
+- Host the drop-in UI separately from this Worker using the assets in
+  `chattia/embed/widget.js`.
 - Configure allowed origins via the `UI_ALLOWED_ORIGINS` environment variable
   (comma-separated list, or `*` for trusted broad access). Pair this with a KV
   table that maps origin → tenant ID to enable per-customer routing.
-- Mount automatically with a `<div data-ultimate-chatbot></div>` placeholder, or
-  call `window.UltimateChatEmbed.mount(element, options)` manually to customize
-  theme, endpoint, or welcome message.
+- Mount the widget with a `<div data-chattia-embed></div>` placeholder, or call
+  `window.ChattiaEmbed.mount(element, options)` manually to customize theme,
+  endpoint, or welcome message.
 - Enforce document prerequisites by checking the tenant record before
   establishing a session. If required PDFs have not been uploaded and approved,
   return a descriptive error so the host site can prompt the customer to supply
