+++ worker/layers/layer4-tiny-ml/README.md
@@ -12,7 +12,7 @@ that enrich conversations without invoking full LLM inference.
 - Emit telemetry to Layer 7 for performance and audit requirements.
 
 ## Setup Checklist
-1. Deploy the Worker `infrastructure/edge-mesh/workers/layer4-tiny-ml`.
+1. Deploy the Worker `edge-mesh/workers/layer4-tiny-ml`.
 2. Bind `HMAC_SECRET` and optional `VECTORIZE_INDEX` or KV namespaces.
 3. Load deterministic ML assets (BM25 tables, embeddings) at publish time,
    scoped by tenant identifiers stored in KV or R2.
